Large update with several pieces:
1. The discrete kalman filter has a new empty implementation.
2. A base implementation of a driver for the MS56xx sensors has been
completed and functions correctly.  The current implementation does not
yet use quantities, and that should be added along with reasonable base
classes for proxies of pressure and temperature.
3. A CRTP based implementation of the LSM6DSx sensor driver has been
implemented up to the point that the matrix operations need to be
replaced with Eigen and the remainder of the configuration functions
should be written (low-pass and high-pass filters, mostly).
4. Performance timers were added to help with performance measurement.
5. An initial HD44870 style LCD driver was added and based on the stream
fuctionality.  Further additions that need to be made include proper
manipulator based cursor positioning, custom characters, read
functionality, support of 40x4 LCDs, optimizations, screen shift, and
backlight control.
6. Additionally there were some changes to other libraries to ensure
correct functionality with the new code and the LPC84x series of
processors.