MPU6050 and lots of small modifications for a new board.
This commit contains a number of fairly small modifications to a large
number of the files that came about as part of the "bringing up" of a
new LPC845 board with an LSM6DS3 and an MS5637.  The bigger change is
the addition of an MPU60x0 driver.  Currently the 60x0 driver is
hardcoded for the lowest range and it does not handle datarate or
lowpass filtering.  It does, however, read from the accelerometer and
gyroscope when driven by an LPC824.  It has not been tested with any
other processor.