Large commit across library Core and utilities updated New files to handle features like Horner's Method for computing Polynomials, accessing the NXP hardware timers, and frameworks for calculating air and atmosphere properties were added.  Currently the air and atmosphere property features are mostly skeletal, but the API should be sufficient for most uses -- at least in the case of the atmosphere.
To support the atmosphere API, a number of units were added to the
library including energy, density and amount units.  In addition, there
were some internal changes to the units system
(dimensions/units/quantities), to make it both safer and more flexible.

There were also a number of random updates to some files to improve
functionality with boards.