Began implementing unit square root.
It has become necessary - to support the NRV (Normal Random Variable)
value type, to have the ability to take the square root of a quantity.
As a result, I also need to be able to take the square root of a
chandra::Dimensions, type as well as a Units type.  The Dimensions type
should be working properly.  It simply divides each dimension by two, static
asserting if any dimension would produce a non-integer result.  The
SqrtUnit is also working on the surface, but the scaling factor is not
correct.  This is going to require calculating the square root of a
std::ratio.  I have identified a possible solution on stack overflow
"Rational Approximation of square root of stdratio at compile time", but
I have not analyized it to see if it is suitable.  In either case, the
current implementation compiles but is functionally incorrect.