Added a compile-time sqrt function to the metaprogramming utils.
The current implementation is sufficient for calculating compile-time
square root of integers and it will successfully handle floating-point,
though as it uses binary search, the accuracy of floating-point is
limited by the maximum iterations (currently set to 250).  For better
overall functionality, a Newton-Raphson iteration should be used for
floating point values while keeping the binary search for integers.