Added a ratio_sqrt implementation to complete the quantity sqrt().
While the quantity sqrt() is now working (after correctly converting
the value through base units -- which would be better to not do, if possible), the tested conversion sqrt(1 yd^2) -> x ft is not entirely accurate, returning 2.99016 instead.  This appears to be an issue steming from the way the unit build util is creating the conversion factors.  Since there is an exact relationship between m, yd, and ft, it should be possible for the factor to, similarly, be exact.  This should transfer to area and volume as well. As such, reconsidering the way the unit build routines work probably makes sense.  In any case, they could stand to be cleaned up.  The simplest solution would be to be able to mark conversions as exact and propogate that through the system, disallowing ratios from being truncated.