Substantial restructuring of the build_units.py script.
Many of the changes to the unit building script were required to allow
for more extensive expression processing for definition of both
conversion factors as well as offset values.  Included in that is making
the different fields allow general Python expressions (so long as they
do not collide with the basic unit definition structure) and adding some
special functions which can be called to do things like exactly convert
floating-point values (in the form of strings), do decade offset, and
directly enter fractions.  Additionally, there were internal processing
changes which improve accuracy of the conversion ratios determined and
which removed duplication in the code - leading to a cleaner
implementation.  Full testing of the units library needs to be
implemented using the Catch2 test framework.  At the very least, every
single base unit should be tested against every non-base unit that
requires a conversion to be manually set for it.

Future work should include caching of errors and simply checking for
more errors in the course of building the units files.  Some errors to
check for could include validating symbol names and validating the
dimension specifications.