Fixed some bugs in the units descriptions.
Electrical Potential (voltage) was defined with the incorrect dimensions
which caused a compilation failure.  Additionally, the scale on the
electronics related units were incorrect as they were not taking the
offset off mass (kg vs. g as the base unit) into account.  It would be
nice to figure out a way to do better checking of all the dimensions and
scales.  Also, it would be handy to simply calculate base scales using
the dimensions of compound units and the scales of base units.