Additional work on the Normal Random Value classes.
Quantity was modified to require scalar operations to be done with
fundamental types so that the multiplication and division operations do
not conflict with the NRV multiplication and division overloads.  This
is a simply way to disambiguate that seems to be reasonable, but it
might make sense to think about this more carefully and - additionally -
add disambiguation to the NRV overloads to make more liberal acceptance
in the quantity overloads possible.