Bugfix and feature addition to HopeRF RFM9x driver.
A bug in the power setting function was fixed so that the power set
function is now working correctly for the RFM9x.  While the chip
supports either RFO or RFBOOST for output, the board only has the
RFBOOST pin connected, so low powers are still unimplemented, though the
current code has a simple boolean variable in the function to allow
turning lower power handling on.

In addition, a simple function to read the number of available bytes was
added, along with a function to set the center frequency of the
transceiver.