Cleaned up baud read function and improved const correctness.
The actual_uclk() read function was not marked const as it should be, so
that was added.  Further testing should be done to ensure that there are
no follow on effects of the change.  Nevertheless, since actual_uclk()
is a read-only function, the change was correct, and so the driver
should be fixed to allow it to be const, if the change does cause
problems.