Added a first implementation of a loopback stream.
This stream currently takes both an echo stream and a
character-at-a-time processor.  Ideally, this needs to allow
construction of simpler loopback streams that take either the echo
stream or the processor, or neither (to provide pure loopback).  The
make function can be overloaded by doing type matching to the Stream
interface to deliniate between the two single argument cases.