Added basic ellipsoid definitions to the geoid header.
The ellipsoid is defined by the Geoid enum class values, and is
currently limited to WGS84 and GRS80.  The gravity classes which are
already written were modified to be based on an Ellipsoid, however, they
still are not properly propogating the geoid definition through, so it
is working entirely with WGS84.  The gravity tests were updated to run a
test on WGS84 and the constructors of the Quantity type were made
constexpr so that quantities could be used in the Ellipsoid class at
compile time.