Implemented ECEF to LLH coordinate conversion.
To get the conversion to work properly and allow the use of the units
library, it was necessary to implement a scalar reduction functionality
as well as make a few tweaks to the quantity library.  Additionally, a
number of functions (atan2(), abs(), cos(), etc.) needed to be added to
the quantity_math.h header to allow for proper calculation with
quantities.  Overall, there are a lot of functions that should be added
to quantity_math.  At the very least, the full set of trigonmetric
functions should be implemented.

For testing to pass the units definitions needed to be regenerated with
a more accurate conversion value between degrees and radians.  When
tested, the algorithm was working correctly, but the results were not
accurate enough to pass the tests.  As such, it probably makes sense to
spend some time looking through the conversion factors to see if any of
the others could be improved.