Many random changes, working toward adding coordinate conversions.
A frame template parameter was added to the matrix vector class that
will be used to cause a compile-time assertion if operations are
attempted on vectors in different reference frames.  Currently, the
frame tag is unused.  It will, however, compile cleanly.

The coordinates header provides a definition for geodetic coordinates
(latitude/longitude/altitude) and geocentric cartesian coordinates
(Earth-Centered/Earth-Fixed), as well as conversions between the two
types.  Additionally, some test cases were added for both the vectors
and quantities.

There were a few other bugfixes in the library.