Added a binary serialization library to allow for easier implementation of binary protocols.
The library is certainly not complete but it should provide basic
serialize/deserialize functionality for unsigned integer values.  I do
need to get it working for, at least, signed integer values.  It would
also be nice to include additional types in the future.  Perhaps
including character strings, floating-point, etc.

Currently, as well, the serialized size is required to be passed in.  If
a value is going to be serialized in its entirety, it would be better to
allow passing without explicit specification of size.  This can be done
with an additional overload of the methods that determines the size of
the value directly.