Got a first version of the localization packet encode/decode implemented.
Currently, it is working well, however, the value mapping code was
truncating and it needs to round to the nearest.  However, making that
change requires also adding saturation to the output to ensure that the
output values remain in a valid range.  To do that, I need to create
compile-time calculation of the minimum and maximum values.  That is not
too difficult, but it will have to wait for tomorrow.

The tests do not pass at the moment for two reasons: 1) the output
saturation is not implemented (causing encoding of the minimum value to
fail) and 2) it is attempting to match the encoded and decoded values
perfectly.  That is never going to work, as data is lost.  Rather than
an exact match, I need to write a function that takes error thresholds
for both the position and velocity values and only checks for errors
that are sub-threshold.